LF-4746 Combine constants and types files
