--- packages/webapp/src/components/Finances/DateRangeSelector/types.ts
@@ -1,29 +1,0 @@
-/*
- *  Copyright 2025 LiteFarm.org
- *  This file is part of LiteFarm.
- *
- *  LiteFarm is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  LiteFarm is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
- */
-
-import { Moment } from 'moment';
-import { DateRangeOptions } from '../../DateRangeSelector/constants';
-
-interface CustomRange {
-  startDate?: string | Moment;
-  endDate?: string | Moment;
-}
-
-export interface DateRangeSelection {
-  option?: DateRangeOptions;
-  startDate?: string | Moment;
-  endDate?: string | Moment;
-  customRange?: CustomRange;
-}
